app/api/requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
azure-cognitiveservices-speech==1.38.0
Babel==2.15.0
beautifulsoup4==4.12.3
bech32==1.2.0
bitsandbytes==0.42.0
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cloudpathlib==0.18.1
cmdstanpy==1.2.4
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cosmpy==0.9.2
cryptography==42.0.8
customtkinter==5.2.2
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.8
darkdetect==0.8.0
dataclasses-json==0.6.7
datasets==3.0.1
db.py==0.5.3
debugpy==1.8.2
decorator==5.1.1
deepface==0.0.91
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
docx2txt==0.8
ecdsa==0.19.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
Flask-Cors==1.10.3
Flask-JWT-Extended==4.6.0
flask-opencv-streamer==1.4
flatbuffers==24.3.25
fonttools==4.52.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.5.4
gdown==5.2.0
gensim==4.3.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
holidays==0.54
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imbalanced-learn==0.13.0
imblearn==0.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaggle==1.6.17
keras==3.3.3
kiwisolver==1.4.5
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.16
langchain-openai==0.1.15
langchain-text-splitters==0.2.1
langcodes==3.4.0
langsmith==0.1.81
language_data==1.2.0
libclang==18.1.1
lightning==2.4.0
lightning-utilities==0.11.8
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.3.2
mongoengine==0.28.2
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nada-ai==0.3.1
nada-dsl==0.5.0
nada-numpy==0.3.2
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nillion-python-helpers==0.2.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
notify2==0.3.1
np_utils==0.6.0
numpy==1.26.4
openai==1.35.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parsial==0.1.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.2
peft==0.13.2
pexpect==4.9.0
phonenumbers==8.13.40
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
prettytable==0.7.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
prophet==1.1.5
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-nillion-client==0.4.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pybars3==0.9.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
PyMeta3==0.5.1
pymongo==4.8.0
PyMsgBox==1.0.9
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyobjc-core==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-Quartz==10.3
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.8.2
PyRect==0.2.0
pyresparser==1.0.6
pyrsistent==0.20.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.4
pytorch-lightning==2.4.0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
ratelim==0.1.6
readme-builder==0.1.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
resume-parser==0.8.4
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
richreports==0.2.0
rpds-py==0.18.1
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.4.3
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.0.1
setuptools==70.0.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
stanio==0.5.1
stemming==1.0.1
streamlit==1.35.0
stripe==10.3.0
sympy==1.12.1
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.16.0
thinc==8.2.5
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tika==2.6.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchmetrics==1.5.1
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
trl==0.11.4
twilio==9.2.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.0
tyro==0.8.13
tzdata==2024.1
ultralytics==8.2.26
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.2
vonage==3.16.0
vonage-jwt==1.1.1
vonage-utils==1.1.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wordcloud==1.9.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
